refactored filesystem, refactored crypto to abstract platform better
