+++ packages/pyright-internal/package.json
@@ -18,7 +18,6 @@
         "@iarna/toml": "2.2.5",
         "@yarnpkg/fslib": "2.5.0-rc.9",
         "@yarnpkg/libzip": "2.2.2-rc.7",
-        "assert": "^2.0.0",
         "chalk": "^4.1.1",
         "chokidar": "^3.5.2",
         "command-line-args": "^5.1.3",
