+++ packages/pyright-internal/src/nodeMain.ts
@@ -1,3 +1,11 @@
+/*
+ * nodeMain.ts
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ *
+ * Provides the main entrypoint to the server when running in Node.
+ */
+
 import { BackgroundAnalysisRunner } from './backgroundAnalysis';
 import { run } from './nodeServer';
 import { PyrightServer } from './server';
